完成了添加区块后广播到整个网络中的功能

+++ ql-block-ledger/src/main/java/org/ql/block/ledger/consensus/PowOfWorkForm.java
@@ -1,5 +1,7 @@
 package org.ql.block.ledger.consensus;
 
+import java.math.BigInteger;
+
 /**
  * Created at 2022/7/3 17:32
  * Author: @Qi Long
@@ -7,6 +9,6 @@ package org.ql.block.ledger.consensus;
  */
 public class PowOfWorkForm {
   public String hash;
-  public int nonce;
+  public BigInteger nonce;
   public String miner;
 }
